# src/main/resources/openapi/metadata-api.yaml
openapi: 3.0.3
info:
  title: Metadata API
  description: API for managing metadata extraction and catalog operations
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/metadata/analyze:
    post:
      summary: Trigger metadata analysis for a directory
      description: |
        Initiates metadata extraction (EXIF data, file hashes) and sidecar generation
        for all supported files in the specified directory.
      operationId: analyzeMetadata
      parameters:
        - name: targetBaseDir
          in: query
          required: true
          description: Directory path to analyze for metadata extraction
          schema:
            type: string
          example: "/Users/anton/Pictures/Imported/2024/2024-01-15"
      responses:
        '202':
          description: Metadata analysis accepted and triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeMetadataResponse'
              example:
                message: "Metadata analysis triggered for directory"
                targetBaseDir: "/Users/anton/Pictures/Imported/2024/2024-01-15"
        '400':
          description: Invalid request (missing or invalid targetBaseDir parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /files:
    get:
      summary: List transferred files with pagination and filtering
      description: |
        Returns a paginated list of transferred files organized by date hierarchy (YYYY/YYYY-MM-DD).
        Supports filtering by date range and directory.
      operationId: listFiles
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page number
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            default: 50
        - name: dateFrom
          in: query
          required: false
          description: Filter files from this date (ISO 8601 format)
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: false
          description: Filter files until this date (ISO 8601 format)
          schema:
            type: string
            format: date
        - name: directory
          in: query
          required: false
          description: Filter files in specific directory path
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedFileListResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /files/{id}/metadata:
    get:
      summary: Get file metadata from sidecar
      description: |
        Retrieves detailed metadata for a specific file including EXIF data and AI analysis results.
      operationId: getFileMetadata
      parameters:
        - name: id
          in: path
          required: true
          description: File identifier
          schema:
            type: string
      responses:
        '200':
          description: File metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadataResponse'
        '404':
          description: File or metadata sidecar not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /files/{id}/thumbnail:
    get:
      summary: Get file thumbnail
      description: |
        Returns a thumbnail image (max 200x200px) for the specified file.
        Only available for image files (JPG, JPEG, PNG).
      operationId: getFileThumbnail
      parameters:
        - name: id
          in: path
          required: true
          description: File identifier
          schema:
            type: string
      responses:
        '200':
          description: Thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: File not found or not an image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /files/search:
    get:
      summary: Search files by various criteria
      description: |
        Searches files by text query (in filenames, AI descriptions, and tags),
        with optional filtering by tags and date range.
      operationId: searchFiles
      parameters:
        - name: q
          in: query
          required: false
          description: Text search query
          schema:
            type: string
        - name: tags
          in: query
          required: false
          description: Comma-separated list of tags to filter by
          schema:
            type: string
        - name: dateFrom
          in: query
          required: false
          description: Filter files from this date (ISO 8601 format)
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: false
          description: Filter files until this date (ISO 8601 format)
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          description: Zero-based page number
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedFileListResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    AnalyzeMetadataResponse:
      type: object
      description: Confirmation response for metadata analysis request
      required:
        - message
        - targetBaseDir
      properties:
        message:
          type: string
          description: Confirmation message indicating analysis was triggered
          example: "Metadata analysis triggered for directory"
        targetBaseDir:
          type: string
          description: Directory path being analyzed
          example: "/Users/anton/Pictures/Imported/2024/2024-01-15"

    FileListResponse:
      type: object
      description: File list item with summary information for file browser
      required:
        - id
        - path
        - filename
        - size
        - transferDate
      properties:
        id:
          type: string
          description: Unique file identifier
          example: "f7c3bc1d-5a2e-4b3f-8c9d-1e2f3a4b5c6d"
        path:
          type: string
          description: Full file path
          example: "/Users/anton/Pictures/Imported/2024/2024-01-15/DSC_2345.JPG"
        filename:
          type: string
          description: File name without path
          example: "DSC_2345.JPG"
        size:
          type: integer
          format: int64
          description: File size in bytes
          example: 4567890
        transferDate:
          type: string
          format: date-time
          description: Date and time when the file was transferred (ISO 8601)
          example: "2024-01-15T10:32:45Z"
        thumbnailUrl:
          type: string
          description: URL to retrieve the file thumbnail (optional, only for images)
          example: "/files/f7c3bc1d-5a2e-4b3f-8c9d-1e2f3a4b5c6d/thumbnail"

    PagedFileListResponse:
      type: object
      description: Paginated response containing file list items
      required:
        - content
        - page
        - size
        - totalPages
        - totalElements
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FileListResponse'
        page:
          type: integer
          description: Current page number (zero-based)
        size:
          type: integer
          description: Number of items per page
        totalPages:
          type: integer
          description: Total number of pages
        totalElements:
          type: integer
          format: int64
          description: Total number of items

    FileMetadataResponse:
      type: object
      description: Detailed file metadata including EXIF and AI analysis
      required:
        - id
        - path
        - filename
        - size
        - transferDate
        - hash
        - hashAlgorithm
      properties:
        id:
          type: string
          description: Unique file identifier
        path:
          type: string
          description: Full file path
        filename:
          type: string
          description: File name without path
        size:
          type: integer
          format: int64
          description: File size in bytes
        transferDate:
          type: string
          format: date-time
          description: Date and time when the file was transferred
        sourcePath:
          type: string
          description: Original source path before transfer
        hash:
          type: string
          description: File hash value
        hashAlgorithm:
          type: string
          description: Hash algorithm used (e.g., MD5, SHA-256)
        exif:
          $ref: '#/components/schemas/ExifData'
        aiAnalysis:
          $ref: '#/components/schemas/AiAnalysisData'

    ExifData:
      type: object
      description: EXIF metadata extracted from image files
      properties:
        cameraModel:
          type: string
          description: Camera model name
        captureDate:
          type: string
          description: Date and time when the photo was captured
        gpsLatitude:
          type: number
          format: double
          description: GPS latitude coordinate
        gpsLongitude:
          type: number
          format: double
          description: GPS longitude coordinate
        focalLength:
          type: string
          description: Focal length (e.g., "50mm")
        aperture:
          type: string
          description: Aperture value (e.g., "f/2.8")
        iso:
          type: integer
          description: ISO sensitivity
        shutterSpeed:
          type: string
          description: Shutter speed (e.g., "1/250")

    AiAnalysisData:
      type: object
      description: AI-generated analysis results
      required:
        - description
        - tags
        - confidence
      properties:
        description:
          type: string
          description: AI-generated description of the image
        tags:
          type: array
          items:
            type: string
          description: AI-generated tags
        confidence:
          type: number
          format: double
          description: Confidence score (0.0 to 1.0)

    ErrorResponse:
      type: object
      description: Error response for validation failures
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
        path:
          type: string
          description: Request path that caused the error